// src/app/shared/model/ref-industry-type-obj.model.ts
export class RefIndustryTypeObj {
    RefIndustryTypeId: number;
    IndustryTypeCode: string;
    IndustryTypeName: string;
    RefIndustryTypeCategoryId: number;
    LbppCode: string;
    IsActive: boolean;
    RowVersion: string;
    constructor() { this.RefIndustryTypeId = 0, this.RowVersion = "" }
}